Update 02-Data.Rmd
Awkward from to construction, it is unclear why it is from kilogram to microgram, if it is an "arbitary level of precision." These do not represent extremes. Maybe it is meant to sound colloquial? "kg" is a little unusual next to "whole" spelled out. It doesn't need to be shortened here, it is being used as a word.

Really, the title of a book should be capitalized, unless this is intentional. Added a comma too, probably.